% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kprime.r
\docType{package}
\name{kprime-package}
\alias{kprime-package}
\title{statistics concerning the Markowitz portfolio}
\description{
Compute PQR functions of the K-prime distribution
}
\note{
This package is maintained as a hobby.
}
\references{

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
