% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kprime}
\alias{kprime}
\alias{pkprime}
\title{K-prime distribution}
\usage{
pkprime(q, v1, v2, a, b, lower_tail = TRUE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{v1}{the degrees of freedom in the numerator chisquare. When
(positive) infinite, we recover a non-central t 
distribution with \code{v2} degrees of freedom and non-centrality
parameter \code{a}, scaled by \code{b}.
This will be recycled against the \code{q,p,n}.}

\item{v2}{the degrees of freedom in the denominator chisquare.
When equal to infinity, we recover the Lambda prime distribution.
This will be recycled against the \code{q,p,n}.}

\item{a}{the non-centrality scaling parameter. When equal to zero,
we recover the (central) t distribution.
This will be recycled against the \code{q,p,n}.}

\item{b}{the scaling parameter.
This will be recycled against the \code{q,p,n}.}

\item{lower_tail}{boolean of whether to compute the lower tail of the
CDF or quantile.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{pkprime} gives the 
distribution function, \code{qkprime} gives the quantile function, 
and \code{rkprime} generates random deviates.

Invalid arguments will result in return value \code{NaN} with a warning.
}
\description{
Distribution, quantile and random variates of the k-prime distribution.
}
\details{
Suppose \eqn{y \sim \chi^2\left(\nu_1\right)}{y ~ x^2(v1)}, and
\eqn{x \sim t \left(\nu_2, a\sqrt{y/\nu_1}/b\right)}{x ~ t(v2,(a/b) sqrt(y/v1))}.
Then the random variable
\deqn{T = b x}{T = b x}
takes a K prime distribution with parameters 
\eqn{\nu_1, \nu_2, a, b}{v1, v2, a, b}. In Lecoutre's terminology,
\eqn{T \sim K'_{\nu_1, \nu_2}\left(a, b\right)}{T ~ K'_v1,v2(a,b)}

Equivalently, we can think of
\deqn{T = \frac{b Z + a \sqrt{\chi^2_{\nu_1} / \nu_1}}{\sqrt{\chi^2_{\nu_2} / \nu_2}}}{T = (bZ + a sqrt(chi2_v1/v1)) / sqrt(chi2_v2/v2)}
where \eqn{Z} is a standard normal, and the normal and the (central) chi-squares are
independent of each other. When \eqn{a=0}{a=0} we recover
a central t distribution; 
when \eqn{\nu_1=\infty}{v1=inf} we recover a rescaled non-central t distribution;
when \eqn{b=0}{b=0}, we get a rescaled square root of a central F
distribution; when \eqn{\nu_2=\infty}{v2=inf}, we recover a 
Lambda prime distribution.
}
\references{
Poitevineau, J., and Lecoutre, B. "Implementing Bayesian predictive procedures: The K-prime and K-square distributions." 
Computational Statistics & Data Analysis 54, no. 3 (2010): 724-731.
\url{https://arxiv.org/abs/1003.4890}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
