% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kprime.r
\name{kprime-NEWS}
\alias{kprime-NEWS}
\title{News for package 'kprime':}
\description{
News for package \sQuote{kprime}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\kprime}{\CRANpkg{kprime}}
}
\section{\kprime{} Initial Version 0.1.0 (2019-06-24) }{

\itemize{
\item first CRAN release.
}
}

